Revise PWM timing architecture.
We now make the long PWM period wider than the longest servo
signal: 2400us. We do this by reloading the timer on every
interrupt. This way we don't need complicated algorithms to
start the servo pulse at the right time, we just wait for the
sync flag and start the servo pulse.

With a 20% duty cycle this should allow us to achieve 350 Hz.