Add PREPROCESSING_MASTER for light-less masters
In case we don't have lights in the chassis we still may want to have
some intelligence in the chassis near the receiver that reads the
servo values from the receivers so that we only have one single
cable to the slave light controller in the body.
We can add just the microcontroller into the receiver for example.
Enabling PREPROCESSING_MASTER achieves this: all servo values
are read and normalized and sent via the UART to an "intelligent
slave".

This slave will then do all the business logic and also contain the
light table to drive the LEDs. This way we can keep the preprocessing
master generic.