Refactoring: make the number of lights that light_mode switches configurable
This allows to define in the makefile how many lights there are.
The LIGHT_MODE_MASK variable needs to be a bit field that has
the bits on for which we want to support a light. The bits
have to be contigous from LSB up.

Examples:
XR311 has Parking and Main beam, which is 2 lights and therefore
LIGHT_MODE_MASK needs to be 0x03

Dingo has Parking, Main beam, Fog lamps and High beam, which
is 4 lights. LIGHT_MODE_MASK is 0x0f