Refactoring: Pre-processor split into separate file; servo-reader created.
To stream-line the code a new preprocessor.asm was created. It is
a very simple program that reads the channels using servo-reader.asm
and outputs them on the serial port.

servo-reader.asm contains now all functions that read the
steering, throttle and ch3 channels from the RC receiver and
normalizes them.

There will also be a uart-reader.asm with the same interface as
servo-reader.asm that will allow to build master.asm with either
functinality reading directly from the RC receiver or reading
from a preprocessor via the UART.