Refactoring: ensured that BANKSEL is applied wherever necessary.
When switching to relocatable assembly I was not aware that the
compiler may assign banks for variables rather arbitrarily.
Incidentally the programs still worked due to luck.
The issue was discovered when I started to use the 16F1825 chip
where Bank 6 has only 48 bytes and it was the first bank the
linker assigned to UDATA.

To stay sane the following programming style was selected:
- Variables for global use (xl/xh/yl/yh/zl/zh/temp/temp+1) are
  allocated in the 16 byte shared memory that is accessible in
  all banks.
- After a call the register bank is unknown and must be set
  implicitely
- It is assumed that the registers PORTx, INTCON, PIRx and for
  TIMER1 are situated in the same bank. This is true for all
  PIC chips that are currently in use.