Remove VPROG pin and Greyscale initialization code.
Turns out that when powering on the TLC5940 has all outputs
set to high. So it is not necessary to initialize the GS registers.
This allows us to free one pin.
This info comes indirectly from TI's Application Report SLVA259
"LC5940 One-Wire Control – Eliminating Microprocessor Control for
Integrated LED Driver".