Phil's Ford: Turn brake lights on when car is in neutral.
Interim solution until we have the brake lights on after going
forward with a timeout, similar to what we have implemented
for reverse already. This is desired for ESC in "rock crawler mode"
where the ESC can go into reverse any time and has a drag brake
when going into neutral.