Split hw_original into 2 files depending on whether UART is read.
With the introduction of preprocessors it is no longer obvious
what a master is and what a slave in terms of hardware configuration.
A master that reads from a preprocessor needs to be initialized
in the same way as a slave.

The only solution I could come up with was to split the files,
which allows us to choose the proper one in the makefile.
A #define could have worked too, but I didn't want to clutter
CFLAGS up too much, and it would be error prone.