Add manual throttle channel reversing instead of automatic.
Two of the light controller customers had issues with the automatic
detection of the throttle channel. It turns out that not everyone
is driving off forward when switching the car on.

The solution is to disable the automatic detection and perfrom
manual configuration like we implemented already for the steering.
Note that this requires both pre-processor and master being
updated, as the auto detection code lies in the pre-processor.