Show warning error if users try to run the configurator source code
If parser.js is not present then we show a large error message, instructing
the user to download the correct file.
It happened that someone downloaded the source code from Github and tried
to load configurator.html, without having installed gulp or built the Light
program assembler. Then simply nothing happens.

To fix this we renamed configurator.html to source.html.

We also upgraded gulp-inline-source and added error messages to the build
process.