Add light program that blinks the output number
Test light program that flashes the output number.
This is helpful to understand which LED is connected to which
output if you can't trace out
the wireing.

Short flashes represent a count of 1, long flashes a count of 5.
OUT0 is always lit while the other outputs perform their flashes.
All outputs are off for 500 ms to serve as synchronization.

Example: long-short-short = 5+1+1 = OUT7