Fix once and for all loading of initialized data
The issue is that we were using _etext to know the ROM address
where the initialization data is stored.
However, if there are multiple ROM sections then we always
have to ensure to move _etext to the ROM section that is right
before the .data section.This is very error prone.

A better solution is to define _data_load and assigning it
to LOADADDR(.data).

It is surprising that almost all vendor linker scripts use
_etext for the initialization data address. Maybe a case of
copying, just like I did ...