Fix servo reader for Spektrum 4210 and 4220 receivers
Those receivers seem to output ST and TH signals at different
frequencies. This caused "flashing" when the throttle was pulled,
and also caused the throttle not to initialize properly when
it was not near around 1500 us.

The changed algorithm passes on the previously read channel instead
of "0".
This means that the channel with the highest frequency dictates
how often new servo data is passed to the light controller, and
the data of the other channels may reflect an earlier state.