Remove erase_mass as it bricks the chip if the bootloader is protected
The DSU can erase the flash, but the User Row protection is still in
place. The Blackmagic probe can no longer write the firmware due
to the protection, but also does not check for those failures.

After erase_mass it is necessary to remove BOOTPROT and flash row
protection. BOOTPROT can not be modified with the standard Blackmagic
probe, it requires a hacked firmware.