Refactor to use common usb descriptors for bootloader and main app
Devices on the market have different USB product ID for the bootloader
and the app, so we copied that design in the beginning.
However, when using WebUSB the user would have to authorize the
same product twice: once in application mode and once in bootloader
mode. This is very inconvenient.

The solution (not verified yet) is to have both the bootloader
and the app use the same USB descriptor, except for the DFU
Interface protocol. This way only one authorization needs to be
given.

On Windows a single driver install should take care of both
bootloader and app mode. Worst case the device is in bootloader
so a PC tool using one of the interfaces that only makes sense
when the app is running will not work.