Implement shorting of light controller power
When in ISP mode, turning the power off causes the MCU to
not reset properly. Most likely there is no brown-out detection
and the MCU is in deep sleep. Once power is reapplied it seems
to be in a weird state.

This is not an issue when the light controller program is running
as it seems to keep the MCU active, draining the power quickly.

The solution is to put a MOSFET and a 100 Ohm resistor accross
the supply rail and short it out when the power is off.
This way everything works reliably.