Make right indicator start flashing immediately
Since quite a long time, the left indicator started to
engage immediately, but the right indicator only after a
short while. After several years someone found out...

The issue was that actually the left indicator was wrong:
BLINK_LEFT_ARMED was supposed to ensure that the steering
wheel is held for a certain amount of time before the indicators
engage. The check was implemented wrong for the left indicators
but correct for the right ones.

Since it is better anyway to have the indicators come one
immediately we can save a bit of code and remove the BLINK_X_ARMED
funcitons