Implement IF statement for light programs
The existing 'skip if' instruction, being negative logic, is
sometimes difficult to reason.
Due to its negative logic the next instruction is not executed
if the statement is true. This often leads to mistakes when
testing for a value range.

We therefore implemented an 'if' statement where the next
instruction is being executed when test expression is true.

Implementation-wise this has been implemented without adding any
new opcodes. Instead the compiler converts the 'if' statement
into a 'skip if' statement.

This works for the following test expressions:
==, !=, <, >, <=, =>, is, not, any, none

It does not work for 'all', as there is no opposite skip if
test statement available.
This means that 'if all forward blink-left' leads to a
compilation error, while 'skip if all forward blink-left' is
perfectly correct.