Add invert_out15s and prefer_all_lights_off features
invert_out15s inverts the switched light output on a low level.
Requested in the German forum to control the enable pin of
an external DC/DC converter

prefer_all_lights_off changes the behaviour of 3 clicks.
When enabled, if the lights are on in any position, then 3-clicks
now toggles the lights to off.
Only when the lights are off, then 3-clicks switches to the highest
light switch position.