Implement program-state-x run-condition
We now have global variables program-state-0 to program-state-5.
We can run light programs when these global variables have a
value other than 0.

This allows us for example to run light programs based on other
conditions.
For example, we can have a "run always" light program that
polls AUX2, and when the value is >50 it sets program-state-0
so that another light program runs.
This triggered light programs can then take over LEDs, without
having to deal with the fact that the LEDs are always taken over
(even when AUX2 is <50 due to "run always")