Optimize light program storage space usage
Originally we reserved space for 25 light programs. So we had
25 *4 bytes in use even if only one light program was defined.

This was not necessary as the firmware already dynamically processed
the start offset, based on the number_of_programs value.

We can save significant space by using number_of_programs instead
of MAX_LIGHT_PROGRAMS (25) when populating the data structure.